Move test code out of the testdata directory.
Test code in the testdata directory breaks the convention used by
certain build system, where "testdata" only contains data such as images
and blobs.

This change moves the test data to a "testgraph" directory, to avoid
breaking such conventions.
