--- graph/testdata/doc.go
@@ -1,2 +1,0 @@
-// Package testdata contains graph test data.
-package testdata // import "github.com/gyuho/goraph/graph/testdata"
