+++ graph/testgraph/doc.go
@@ -1,0 +1,2 @@
+// Package testgraph contains graph test data.
+package testgraph // import "github.com/gyuho/goraph/graph/testgraph"
